avanzo en el front, se puede, crear cuenta, logear, mandar solicitud, revisar mis solicitudes, revisar y evaluar solicitudes de todos, aprobar o rechazar solicitudes, realizar una simulacion de un credito viendo la cuota mensual, queda PENDIENTE poder ver los costes totales de una solicitud propia, no esta hecho ni el backend

+++ Backend/prestabank-backend/src/main/java/edu/mtisw/prestabank_backend/Entity/LoanEntity.java
@@ -38,7 +38,7 @@ public class LoanEntity {
     private ArrayList<Integer> evalue; //   Listado que indica automaticamente las cosas 0=rechazado, 1=aprobado, 2=pendiente. 3=Requiere otra revision [R1, R2, R3, R4, R6, R7]
 
 
-    public LoanEntity(Long idUser, int type, int yearInterest, int maxDuration, int maxLoan, int income, int veteran, int totaldebt, double monthlyPayment, double loanAmount, double monthlyInteresRate, int totalPayments, ArrayList<Integer> saving, ArrayList<String> papers, ArrayList<Integer> evalue) {
+    public LoanEntity(Long idUser, int type, double yearInterest, int maxDuration, int maxLoan, int income, int veteran, int totaldebt, double monthlyPayment, double loanAmount, double monthlyInteresRate, int totalPayments, ArrayList<Integer> saving, ArrayList<String> papers, ArrayList<Integer> evalue) {
         this.idUser = idUser;
         this.status = 1;
         this.type = type;
